CSV-Converter + Create Article Fix
Added a default keywords value when creating an article called "anti-crash". MySQL Expects a deprecated keyword to be added when creating an article. This dummy text should provide it with said default value. Also added a csv to json converter in the background to make converting stats easier. Preparation for future json -> stats upload in article creation